feat: wizard
